move slashing and tally to keeper
