+++ x/oracle/keeper/slash_test.go
@@ -1,4 +1,4 @@
-package oracle
+package keeper
 
 import (
 	"testing"
