+++ x/oracle/keeper/tally.go
@@ -1,10 +1,9 @@
-package oracle
+package keeper
 
 import (
 	"sort"
 
 	sdk "github.com/cosmos/cosmos-sdk/types"
-	"github.com/peggyjv/sommelier/x/oracle/keeper"
 	"github.com/peggyjv/sommelier/x/oracle/types"
 )
